Split six files back under the line limit
Each was one module holding several, and each split along a seam that
already existed rather than by line count:

* main_widget.py keeps registration and layout; the screenshot, image
  detection, recording and script tabs move to their own modules.
* webrtc_dialogs.py sheds known-host storage and verification.
* remote_desktop/host.py sheds viewer access control, host capture and
  the per-client handler.
* webrtc_host.py sheds authentication and media negotiation.
* mcp_server/server.py sheds the wire protocol and client-initiated
  requests; http_transport takes _notification_message from _protocol.
* flow_control.py sheds the data-producing commands (ocr / pdf / shell
  / sql / otp to var, assert_duration).

No behaviour change: the moves are re-exported where callers expect
them, and the tests follow the symbols to their new homes.
