+++ test/unit_test/headless/test_remote_desktop_pending_viewer.py
@@ -4,9 +4,8 @@ import time
 
 import pytest
 
-from je_auto_control.utils.remote_desktop.host import (
-    PendingViewer, RemoteDesktopHost,
-)
+from je_auto_control.utils.remote_desktop.host import RemoteDesktopHost
+from je_auto_control.utils.remote_desktop.host_access import PendingViewer
 from je_auto_control.utils.remote_desktop.protocol import (
     AuthenticationError,
 )
